+++ lib/economy_embeds.py
