+++ lib/embeds.py
