+++ lib/resources.py
